ui: global state
