+++ frontend/src/components/Deployer.tsx
@@ -1,46 +1,15 @@
-import React, { useEffect, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import Urbit from '@urbit/http-api';
+import { GlobalState, GlobalStateContext } from './Global';
 
-interface DeskMetadata {
-  desk: string;
-}
 
 const api = new Urbit('', '', window.desk);
 api.ship = window.ship;
 
 export function Deployer() {
-  // @ts-ignore
-  const [desks, setDesks] = useState<Array<string>>();
 
-  useEffect(() => {
-    async function init() {
+  const { desks } = useContext(GlobalStateContext);
 
-      api.subscribe({
-        app: "vita-deploy",
-        path: "/frontend",
-        event: handleSub,
-        quit: () => alert("lost connection to your urbit. please refresh"),
-        err: (e) => console.log("error", e),
-      }).then((subscriptionId) => {
-        //
-        window.addEventListener("beforeunload", () => {
-          api.unsubscribe(subscriptionId);
-          api.delete();
-        });
-      });
-    }
-
-    init();
-  }, []);
-
-  function handleSub(data: any) {
-    console.log('got update', data)
-    const desks = 'desks'
-    if (data[desks] !== undefined) {
-      const newDesks: Array<string> = data[desks]
-      setDesks(newDesks);
-    }
-  }
 
   const invalidDeskNames = ['base', 'landscape', 'groups', 'talk', 'realm', 'vita', 'lemur', 'pals', 'radio', 'rumors', 'portal']
   function newDesk(deskName: string) {
@@ -87,11 +56,10 @@ export function Deployer() {
           </div>
         </div>
 
-
         <div>
           <ul>
             {/* created desks */}
-            {desks?.map((desk: string) => (
+            {desks.map((desk: string) => (
               <li key={desk} >
                 <div
                   style={{
