+++ frontend/src/components/Global.tsx
@@ -1,0 +1,58 @@
+import Urbit from '@urbit/http-api';
+import React, { createContext, useEffect, useState } from 'react';
+
+export interface GlobalState {
+  desks: Array<string>;
+}
+
+const buntGlobalState = {
+  desks: [],
+}
+
+export const GlobalStateContext = createContext<GlobalState>(buntGlobalState);
+
+const api = new Urbit('', '', window.desk);
+api.ship = window.ship;
+
+export const GlobalStateProvider = ({ children }: { children: React.ReactNode }) => {
+  const [state, setState] = useState<GlobalState>(buntGlobalState);
+
+  useEffect(() => {
+    async function init() {
+
+      api.subscribe({
+        app: "vita-deploy",
+        path: "/frontend",
+        event: handleSub,
+        quit: () => alert("lost connection to your urbit. please refresh"),
+        err: (e) => console.log("urbit http-api error:", e),
+      }).then((subscriptionId) => {
+        //
+        window.addEventListener("beforeunload", () => {
+          api.unsubscribe(subscriptionId);
+          api.delete();
+        });
+      });
+    }
+
+    init();
+  }, []);
+
+  function handleSub(data: any) {
+    console.log('got update', data)
+    const desks = 'desks'
+    if (data[desks] !== undefined) {
+      const newDesks: Array<string> = data[desks]
+      let newState = { ...state };
+      newState.desks = newDesks;
+      setState(newState);
+    }
+  }
+
+
+  return (
+    <GlobalStateContext.Provider value={state}>
+      {children}
+    </GlobalStateContext.Provider>
+  );
+};
